feat : 로또(자동)
추가내용
* 당첨정보에서는 랭크 매칭 정보만 리턴을 하고 결과자체를 로또최종등수정보 객체에게 주입하여 결과값을 가지도록 수정함.
