+++ src/main/java/lotto/domain/LottoResult.java
@@ -10,13 +10,17 @@ public class LottoResult {
 
   private final Map<Rank, List<Lotto>> categoriesRank = new LinkedHashMap<>();
 
-  public LottoResult() {
+  private WinLottoInfo winLottoInfo;
+
+  public LottoResult(final WinLottoInfo winLottoInfo) {
     categoriesRank.put(Rank.FIFTH, new ArrayList<>());
     categoriesRank.put(Rank.FOURTH, new ArrayList<>());
     categoriesRank.put(Rank.THIRD, new ArrayList<>());
     categoriesRank.put(Rank.SECOND, new ArrayList<>());
     categoriesRank.put(Rank.FIRST, new ArrayList<>());
     categoriesRank.put(Rank.MISS, new ArrayList<>());
+
+    this.winLottoInfo = winLottoInfo;
   }
 
   public Map<Rank, List<Lotto>> getCategoriesRank() {
@@ -31,4 +35,17 @@ public class LottoResult {
     return categoriesRank.get(matchCountForRank);
   }
 
+  public LottoResult matchLottoInfo(Lotteries lotteries) {
+
+    for (Lotto lotto : lotteries.getLottos()) {
+      getLottosByRank(lotto).add(lotto);
+    }
+
+    getCategoriesRank().remove(Rank.MISS);
+    return this;
+  }
+
+  private List<Lotto> getLottosByRank(final Lotto lotto) {
+    return getMatchLottos(winLottoInfo.getMatchCountForRank(lotto));
+  }
 }
